Amélioration de la structure et ajouts de ressources

+++ app/Http/Controllers/ProjectController.php
@@ -4,47 +4,57 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Response\JsonResponse;
+use Validator;
 use App\Project;
 use App\Status;
 use App\User_projects;
+use App\Http\Resources\Project as ProjectResource;
 class ProjectController extends Controller
 {
+  /**
+  * Returns all existing projects with details
+  * @return array
+  */
   public static function showAll(){
     $projects = Project::getAll();
     $response = new JsonResponse;
-    $response->setData($projects);
-    return $response->throw();
+    return $response->setData($projects);
   }
 
 
 
-
+  /**
+  * Returns the specified project with details
+  *@param int $id
+  * @return array
+  */
   public static function get($id){
+    
+    $project = new ProjectResource(Project::find($id));
     $response = new JsonResponse;
     if(!is_null(Project::find($id))){
-      $project = Project::getWithDetails($id);
-      $response->setData($project);
+      return $response->setData($project);
     }else{
-      $response->addErrors(["Project ".$id." doesn't exist."]);
+      return $response->addErrors(["Project ".$id." doesn't exist."]);
     }
-    return $response->throw();
   }
 
 
-
+  /**
+  * Add this project
+  * @param string $label
+  * @param int $progress
+  */
   public static function add(Request $request){
-    $request->validate([
+    $validator = Validator::make($request->all(),[
       'label' => 'string|required',
-      'progress' => 'integer'
+      'progress' => 'integer|required'
     ]);
-    if($request->filled('progress')){
-      if($request->progress <=100 && $request->progress >=0){
-        $progress = $request->progress;
-      }
-      else{
-        $progress = 0;
-      }
+    $response = new JsonResponse;
+    if($validator->fails()){
+      return $response->addErrors(array($validator->errors()));
     }
+
     if($request->filled('status_id')){
       $status_id = $request->status_id;
     }else{
@@ -55,45 +65,65 @@ class ProjectController extends Controller
       'progress' => $progress,
       'status_id' =>$status_id
     ]);
-
-    $response = new JsonResponse;
-    $response->setData($project);
-    return $response->throw();
+    $formattedproject = new ProjectResource($project);
+    return $response->setData($project);
   }
 
-
+  /**
+  * Update a specific project
+  * @param int $id
+  * @param int $progress
+  * @param string $label
+  * @return array
+  */
   public static function update(Request $request){
-    $request->validate([
+    $validator = Validator::make($request->all(),[
       'id' => 'integer|exists:projects,id',
-      'progress' => 'integer',
+      'progress' => 'integer|required',
       'label' => 'string|required'
     ]);
+    $response = new JsonResponse;
+    if($validator->fails()){
+      return $response->addErrors(array($validator->errors()));
+    }
 
     Project::where('id', $request->id)->update([
       'label' => $request->label,
       'progress' => $request->progress,
       'status_id' => $request->status_id
     ]);
-    $project = Project::getWithDetails($request->id);
-    $project->status = Status::where('id', $project->status_id)->first(['id','label']);
-    unset($project->status_id);
-    $response = new JsonResponse;
-    $response->setData($project);
-    return $response->throw();
+    $project = new ProjectResource(Project::find($request->id));
+    return $response->setData($project);
   }
 
+
+
+
+  /**
+  *Detelete a project
+  * @param int id
+  * @return array
+  */
   public static function delete(Request $request){
-    $request->validate([
-      'id' => 'integer|exists:projects,id'
+    $validator = Validator::make($request->all(),[
+      'id' => 'required|integer|exists:projects,id'
     ]);
-    Project::destroy($request->id);
     $response = new JsonResponse;
-    $response->setMessage('Done.');
-    return $response->throw();
+    if($validator->fails()){
+      return $response->addErrors(array($validator->errors()));
+    }
+    Project::destroy($request->id);
+    User_projects::where('project_id', $request->id)->delete();
+
+    return $response->setMessage('Done.');
   }
 
 
-
+  /**
+  * @param int $users_id
+  * @param int project_id
+  * @return array
+  */
   public static function assign(Request $request){
     $request->validate([
       'users_id' => 'exists:users,id|required',
@@ -102,14 +132,13 @@ class ProjectController extends Controller
     $users = $request->users_id;
     User_projects::where('project_id', $request->project_id)->delete();
     foreach ($users as $user) {
-      User_projects::create([
+      User_projects::firstOrCreate([
         'user_id' => $user,
         'project_id' => $request->project_id
       ]);
     }
     $response = new JsonResponse;
-    $response->setData(Project::getWithDetails($request->project_id));
-    return $response->throw();
+    return $response->setData(Project::getWithDetails($request->project_id));
   }
 
 }
